+++ src/main/kotlin/com/team2/server/party/infrastructure/persistence/RealtimeParticipantProfileRepository.kt
@@ -1,10 +1,9 @@
-package com.team2.server.party.repository
+package com.team2.server.party.infrastructure.persistence
 
 import com.team2.server.party.domain.entity.Participant
 import com.team2.server.party.domain.entity.RealtimeParticipantProfile
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.data.jpa.repository.Modifying
-import org.springframework.transaction.annotation.Transactional
 
 interface RealtimeParticipantProfileRepository : JpaRepository<RealtimeParticipantProfile, Long> {
     fun findByParticipant(participant: Participant): RealtimeParticipantProfile?
@@ -12,6 +11,5 @@ interface RealtimeParticipantProfileRepository : JpaRepository<RealtimeParticipa
     fun findByParticipantToken(participantToken: String): RealtimeParticipantProfile?
 
     @Modifying
-    @Transactional
     fun deleteAllByParticipantIdIn(participantIds: List<Long>)
 }
