+++ src/main/kotlin/com/team2/server/party/service/ParticipantService.kt
@@ -3,7 +3,7 @@ package com.team2.server.party.service
 import com.team2.server.common.exception.isConstraintViolation
 import com.team2.server.party.domain.entity.Participant
 import com.team2.server.party.domain.entity.Party
-import com.team2.server.party.repository.ParticipantRepository
+import com.team2.server.party.infrastructure.persistence.ParticipantRepository
 import com.team2.server.user.entity.User
 import org.springframework.dao.DataIntegrityViolationException
 import org.springframework.stereotype.Service
